split giant shaped array file into multiple files
