+++ Package.swift
@@ -8,8 +8,8 @@ let package = Package(
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.
         .library(
-            name: "shaped-array",
-            targets: ["shaped-array"]),
+            name: "ShapedArray",
+            targets: ["ShapedArray"]),
     ],
     dependencies: [
         // Dependencies declare other packages that this package depends on.
@@ -19,10 +19,10 @@ let package = Package(
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
         // Targets can depend on other targets in this package, and on products in packages this package depends on.
         .target(
-            name: "shaped-array",
+            name: "ShapedArray",
             dependencies: []),
         .testTarget(
-            name: "shaped-arrayTests",
-            dependencies: ["shaped-array"]),
+            name: "ShapedArrayTests",
+            dependencies: ["ShapedArray"]),
     ]
 )
