+++ Tests/ShapedArrayTests/ShapedArrayTests.swift
@@ -1,7 +1,7 @@
 import XCTest
-@testable import shaped_array
+@testable import ShapedArray
 
-final class shaped_arrayTests: XCTestCase {
+final class ShapedArrayTests: XCTestCase {
     func testExample() throws {
         // This is an example of a functional test case.
         // Use XCTAssert and related functions to verify your tests produce the correct
